Enum mcaptcha::errors::ServiceError

pub enum ServiceError {

    InternalServerError,
    ClosedForRegistration,
    NotAnEmail,
    NotAUrl,
    WrongPassword,
    UsernameNotFound,
    AccountNotFound,
    ProfainityError,
    BlacklistError,
    UsernameCaseMappedError,
    PasswordTooShort,
    PasswordTooLong,
    PasswordsDontMatch,
    UsernameTaken,
    EmailTaken,
    UnableToSendEmail(SmtpErrorWrapper),
    TokenNotFound,
    CaptchaError(CaptchaError),

}

Variants

InternalServerError

ClosedForRegistration

NotAnEmail

NotAUrl

WrongPassword

UsernameNotFound

AccountNotFound

ProfainityError

when the value passed contains profainity

BlacklistError

when the value passed contains blacklisted words see blacklist

UsernameCaseMappedError

when the value passed contains characters not present in UsernameCaseMapped profile

PasswordTooShort

PasswordTooLong

PasswordsDontMatch

UsernameTaken

when the a username is already taken

EmailTaken

email is already taken

UnableToSendEmail(SmtpErrorWrapper)

Unable to send email

Tuple Fields of UnableToSendEmail

0: SmtpErrorWrapper

TokenNotFound

when the a token name is already taken token not found

CaptchaError(CaptchaError)

Tuple Fields of CaptchaError

0: CaptchaError